+++ src/Skylark/Manage/Storage/StorageManage.cs
@@ -16,7 +16,7 @@ namespace Skylark.Manage
         /// <summary>
         /// 
         /// </summary>
-        public const decimal Value = 10;
+        public const double Value = 10;
 
         /// <summary>
         /// 
@@ -54,156 +54,156 @@ namespace Skylark.Manage
         /// <summary>
         /// 
         /// </summary>
-        public static readonly Dictionary<EST, Dictionary<EST, decimal>> Converter = new()
+        public static readonly Dictionary<EST, Dictionary<EST, double>> Converter = new()
         {
             {
                 EST.Bit, new()
                 {
-                    { EST.Bit, 1M },
-                    { EST.Byte, 0.125M },
-                    { EST.Kilobyte, 0.000125M },
-                    { EST.Megabyte, 0.000000125M },
-                    { EST.Gigabyte, 0.000000000125M },
-                    { EST.Terabyte, 0.000000000000125M },
-                    { EST.Petabyte, 0.000000000000000125M },
-                    { EST.Exabyte, 0.000000000000000000125M },
-                    { EST.Zetabyte, 0.000000000000000000000125M },
-                    { EST.Yottabyte, 0.000000000000000000000000125M }
+                    { EST.Bit, 1d },
+                    { EST.Byte, 0.125d },
+                    { EST.Kilobyte, 0.000125d },
+                    { EST.Megabyte, 0.000000125d },
+                    { EST.Gigabyte, 0.000000000125d },
+                    { EST.Terabyte, 0.000000000000125d },
+                    { EST.Petabyte, 0.000000000000000125d },
+                    { EST.Exabyte, 0.000000000000000000125d },
+                    { EST.Zetabyte, 0.000000000000000000000125d },
+                    { EST.Yottabyte, 0.000000000000000000000000125d }
                 }
             },
             {
                 EST.Byte, new()
                 {
-                    { EST.Bit, 8M },
-                    { EST.Byte, 1M },
-                    { EST.Kilobyte, 0.001M },
-                    { EST.Megabyte, 0.0000010M },
-                    { EST.Gigabyte, 0.0000000010M },
-                    { EST.Terabyte, 0.0000000000010M },
-                    { EST.Petabyte, 0.0000000000000010M },
-                    { EST.Exabyte, 0.0000000000000000010M },
-                    { EST.Zetabyte, 0.0000000000000000000010M },
-                    { EST.Yottabyte, 0.0000000000000000000000010M }
+                    { EST.Bit, 8d },
+                    { EST.Byte, 1d },
+                    { EST.Kilobyte, 0.001d },
+                    { EST.Megabyte, 0.0000010d },
+                    { EST.Gigabyte, 0.0000000010d },
+                    { EST.Terabyte, 0.0000000000010d },
+                    { EST.Petabyte, 0.0000000000000010d },
+                    { EST.Exabyte, 0.0000000000000000010d },
+                    { EST.Zetabyte, 0.0000000000000000000010d },
+                    { EST.Yottabyte, 0.0000000000000000000000010d }
                 }
             },
             {
                 EST.Kilobyte, new()
                 {
-                    { EST.Bit, 8000M },
-                    { EST.Byte, 1000M },
-                    { EST.Kilobyte, 1M },
-                    { EST.Megabyte, 0.001M },
-                    { EST.Gigabyte, 0.000001M },
-                    { EST.Terabyte, 0.000000001M },
-                    { EST.Petabyte, 0.000000000001M },
-                    { EST.Exabyte, 0.000000000000001M },
-                    { EST.Zetabyte, 0.000000000000000001M },
-                    { EST.Yottabyte, 0.000000000000000000001M }
+                    { EST.Bit, 8000d },
+                    { EST.Byte, 1000d },
+                    { EST.Kilobyte, 1d },
+                    { EST.Megabyte, 0.001d },
+                    { EST.Gigabyte, 0.000001d },
+                    { EST.Terabyte, 0.000000001d },
+                    { EST.Petabyte, 0.000000000001d },
+                    { EST.Exabyte, 0.000000000000001d },
+                    { EST.Zetabyte, 0.000000000000000001d },
+                    { EST.Yottabyte, 0.000000000000000000001d }
                 }
             },
             {
                 EST.Megabyte, new()
                 {
-                    { EST.Bit, 8000000M },
-                    { EST.Byte, 1000000M },
-                    { EST.Kilobyte, 1000M },
-                    { EST.Megabyte, 1M },
-                    { EST.Gigabyte, 0.001M },
-                    { EST.Terabyte, 0.000001M },
-                    { EST.Petabyte, 0.000000001M },
-                    { EST.Exabyte, 0.000000000001M },
-                    { EST.Zetabyte, 0.000000000000001M },
-                    { EST.Yottabyte, 0.000000000000000001M }
+                    { EST.Bit, 8000000d },
+                    { EST.Byte, 1000000d },
+                    { EST.Kilobyte, 1000d },
+                    { EST.Megabyte, 1d },
+                    { EST.Gigabyte, 0.001d },
+                    { EST.Terabyte, 0.000001d },
+                    { EST.Petabyte, 0.000000001d },
+                    { EST.Exabyte, 0.000000000001d },
+                    { EST.Zetabyte, 0.000000000000001d },
+                    { EST.Yottabyte, 0.000000000000000001d }
                 }
             },
             {
                 EST.Gigabyte, new()
                 {
-                    { EST.Bit, 8000000000M },
-                    { EST.Byte, 1000000000M },
-                    { EST.Kilobyte, 1000000M },
-                    { EST.Megabyte, 1000M },
-                    { EST.Gigabyte, 1M },
-                    { EST.Terabyte, 0.001M },
-                    { EST.Petabyte, 0.000001M },
-                    { EST.Exabyte, 0.000000001M },
-                    { EST.Zetabyte, 0.000000000001M },
-                    { EST.Yottabyte, 0.000000000000001M }
+                    { EST.Bit, 8000000000d },
+                    { EST.Byte, 1000000000d },
+                    { EST.Kilobyte, 1000000d },
+                    { EST.Megabyte, 1000d },
+                    { EST.Gigabyte, 1d },
+                    { EST.Terabyte, 0.001d },
+                    { EST.Petabyte, 0.000001d },
+                    { EST.Exabyte, 0.000000001d },
+                    { EST.Zetabyte, 0.000000000001d },
+                    { EST.Yottabyte, 0.000000000000001d }
                 }
             },
             {
                 EST.Terabyte, new()
                 {
-                    { EST.Bit, 8000000000000M },
-                    { EST.Byte, 1000000000000M },
-                    { EST.Kilobyte, 1000000000M },
-                    { EST.Megabyte, 1000000M },
-                    { EST.Gigabyte, 1000M },
-                    { EST.Terabyte, 1M },
-                    { EST.Petabyte, 0.001M },
-                    { EST.Exabyte, 0.000001M },
-                    { EST.Zetabyte, 0.000000001M },
-                    { EST.Yottabyte, 0.000000000001M }
+                    { EST.Bit, 8000000000000d },
+                    { EST.Byte, 1000000000000d },
+                    { EST.Kilobyte, 1000000000d },
+                    { EST.Megabyte, 1000000d },
+                    { EST.Gigabyte, 1000d },
+                    { EST.Terabyte, 1d },
+                    { EST.Petabyte, 0.001d },
+                    { EST.Exabyte, 0.000001d },
+                    { EST.Zetabyte, 0.000000001d },
+                    { EST.Yottabyte, 0.000000000001d }
                 }
             },
             {
                 EST.Petabyte, new()
                 {
-                    { EST.Bit, 8000000000000000M },
-                    { EST.Byte, 1000000000000000M },
-                    { EST.Kilobyte, 1000000000000M },
-                    { EST.Megabyte, 1000000000M },
-                    { EST.Gigabyte, 1000000M },
-                    { EST.Terabyte, 1000M },
-                    { EST.Petabyte, 1M },
-                    { EST.Exabyte, 0.001M },
-                    { EST.Zetabyte, 0.000001M },
-                    { EST.Yottabyte, 0.000000001M }
+                    { EST.Bit, 8000000000000000d },
+                    { EST.Byte, 1000000000000000d },
+                    { EST.Kilobyte, 1000000000000d },
+                    { EST.Megabyte, 1000000000d },
+                    { EST.Gigabyte, 1000000d },
+                    { EST.Terabyte, 1000d },
+                    { EST.Petabyte, 1d },
+                    { EST.Exabyte, 0.001d },
+                    { EST.Zetabyte, 0.000001d },
+                    { EST.Yottabyte, 0.000000001d }
                 }
             },
             {
                 EST.Exabyte, new()
                 {
-                    { EST.Bit, 8000000000000000000M },
-                    { EST.Byte, 1000000000000000000M },
-                    { EST.Kilobyte, 1000000000000000M },
-                    { EST.Megabyte, 1000000000000M },
-                    { EST.Gigabyte, 1000000000M },
-                    { EST.Terabyte, 1000000M },
-                    { EST.Petabyte, 1000M },
-                    { EST.Exabyte, 1M },
-                    { EST.Zetabyte, 0.001M },
-                    { EST.Yottabyte, 0.000001M }
+                    { EST.Bit, 8000000000000000000d },
+                    { EST.Byte, 1000000000000000000d },
+                    { EST.Kilobyte, 1000000000000000d },
+                    { EST.Megabyte, 1000000000000d },
+                    { EST.Gigabyte, 1000000000d },
+                    { EST.Terabyte, 1000000d },
+                    { EST.Petabyte, 1000d },
+                    { EST.Exabyte, 1d },
+                    { EST.Zetabyte, 0.001d },
+                    { EST.Yottabyte, 0.000001d }
                 }
             },
             {
                 EST.Zetabyte, new()
                 {
-                    { EST.Bit, 8000000000000000000000M },
-                    { EST.Byte, 1000000000000000000000M },
-                    { EST.Kilobyte, 1000000000000000000M },
-                    { EST.Megabyte, 1000000000000000M },
-                    { EST.Gigabyte, 1000000000000M },
-                    { EST.Terabyte, 1000000000M },
-                    { EST.Petabyte, 1000000M },
-                    { EST.Exabyte, 1000M },
-                    { EST.Zetabyte, 1M },
-                    { EST.Yottabyte, 0.001M }
+                    { EST.Bit, 8000000000000000000000d },
+                    { EST.Byte, 1000000000000000000000d },
+                    { EST.Kilobyte, 1000000000000000000d },
+                    { EST.Megabyte, 1000000000000000d },
+                    { EST.Gigabyte, 1000000000000d },
+                    { EST.Terabyte, 1000000000d },
+                    { EST.Petabyte, 1000000d },
+                    { EST.Exabyte, 1000d },
+                    { EST.Zetabyte, 1d },
+                    { EST.Yottabyte, 0.001d }
                 }
             },
             {
                 EST.Yottabyte, new()
                 {
-                    { EST.Bit, 8000000000000000000000000M },
-                    { EST.Byte, 1000000000000000000000000M },
-                    { EST.Kilobyte, 1000000000000000000000M },
-                    { EST.Megabyte, 1000000000000000000M },
-                    { EST.Gigabyte, 1000000000000000M },
-                    { EST.Terabyte, 1000000000000M },
-                    { EST.Petabyte, 1000000000M },
-                    { EST.Exabyte, 1000000M },
-                    { EST.Zetabyte, 1000M },
-                    { EST.Yottabyte, 1M }
+                    { EST.Bit, 8000000000000000000000000d },
+                    { EST.Byte, 1000000000000000000000000d },
+                    { EST.Kilobyte, 1000000000000000000000d },
+                    { EST.Megabyte, 1000000000000000000d },
+                    { EST.Gigabyte, 1000000000000000d },
+                    { EST.Terabyte, 1000000000000d },
+                    { EST.Petabyte, 1000000000d },
+                    { EST.Exabyte, 1000000d },
+                    { EST.Zetabyte, 1000d },
+                    { EST.Yottabyte, 1d }
                 }
             }
         };
